run: There is always some space for the cat!
Actually this is to make an additional, almost empty line to appear between the curl cmd line and its output -- it is to make the tracing output more friendly to a human eye.

And because we pipe output of the curl, we need another option, `-s` to keep it silent.